the struct was split
